Moved 5 funcs to `cotc.py`: get_warring_clan, get_other_clan, get_other_clan_relation, change_clan_relations, change_clan_reputation
