+++ scripts/clan_package/cotc.py
@@ -1,0 +1,68 @@
+from scripts.game_structure import game
+
+
+def get_warring_clan():
+    """
+    returns enemy clan if a war is currently ongoing
+    """
+    enemy_clan = None
+    if game.clan.war.get("at_war", False):
+        for other_clan in game.clan.all_other_clans:
+            if other_clan.name == game.clan.war["enemy"]:
+                enemy_clan = other_clan
+
+    return enemy_clan
+
+
+def get_other_clan(clan_name):
+    """
+    returns the clan object of given clan name
+    """
+    for clan in game.clan.all_other_clans:
+        if clan.name == clan_name:
+            return clan
+
+
+def get_other_clan_relation(relation):
+    """
+    converts int value into string relation and returns string: "hostile", "neutral", or "ally"
+    :param relation: the other_clan.relations value
+    """
+
+    if int(relation) >= 17:
+        return "ally"
+    elif 7 < int(relation) < 17:
+        return "neutral"
+    elif int(relation) <= 7:
+        return "hostile"
+
+
+def change_clan_relations(other_clan, difference):
+    """
+    will change the Clan's relation with other clans according to the difference parameter.
+    """
+    # grab the clan that has been indicated
+    other_clan = other_clan
+    # grab the relation value for that clan
+    y = game.clan.all_other_clans.index(other_clan)
+    clan_relations = int(game.clan.all_other_clans[y].relations)
+    # change the value
+    clan_relations += difference
+    # making sure it doesn't exceed the bounds
+    if clan_relations > 30:
+        clan_relations = 30
+    elif clan_relations < 0:
+        clan_relations = 0
+    # setting it in the Clan save
+    game.clan.all_other_clans[y].relations = clan_relations
+
+
+def change_clan_reputation(difference):
+    """
+    will change the Clan's reputation with outsider cats according to the difference parameter.
+    """
+    game.clan.reputation += difference
+    if game.clan.reputation < 0:
+        game.clan.reputation = 0  # clamp to 0
+    elif game.clan.reputation > 100:
+        game.clan.reputation = 100  # clamp to 100
