+++ scripts/utility.py
@@ -57,19 +57,6 @@ if TYPE_CHECKING:
 # ---------------------------------------------------------------------------- #
 
 
-def get_warring_clan():
-    """
-    returns enemy clan if a war is currently ongoing
-    """
-    enemy_clan = None
-    if game.clan.war.get("at_war", False):
-        for other_clan in game.clan.all_other_clans:
-            if other_clan.name == game.clan.war["enemy"]:
-                enemy_clan = other_clan
-
-    return enemy_clan
-
-
 # ---------------------------------------------------------------------------- #
 #                          Handling Outside Factors                            #
 # ---------------------------------------------------------------------------- #
@@ -94,37 +81,6 @@ def get_current_season():
     game.clan.current_season = constants.SEASON_CALENDAR[index]
 
     return game.clan.current_season
-
-
-def change_clan_reputation(difference):
-    """
-    will change the Clan's reputation with outsider cats according to the difference parameter.
-    """
-    game.clan.reputation += difference
-    if game.clan.reputation < 0:
-        game.clan.reputation = 0  # clamp to 0
-    elif game.clan.reputation > 100:
-        game.clan.reputation = 100  # clamp to 100
-
-
-def change_clan_relations(other_clan, difference):
-    """
-    will change the Clan's relation with other clans according to the difference parameter.
-    """
-    # grab the clan that has been indicated
-    other_clan = other_clan
-    # grab the relation value for that clan
-    y = game.clan.all_other_clans.index(other_clan)
-    clan_relations = int(game.clan.all_other_clans[y].relations)
-    # change the value
-    clan_relations += difference
-    # making sure it doesn't exceed the bounds
-    if clan_relations > 30:
-        clan_relations = 30
-    elif clan_relations < 0:
-        clan_relations = 0
-    # setting it in the Clan save
-    game.clan.all_other_clans[y].relations = clan_relations
 
 
 def create_new_cat_block(
@@ -553,15 +509,6 @@ def create_new_cat_block(
             n_c.create_inheritance_new_cat()
 
     return new_cats
-
-
-def get_other_clan(clan_name):
-    """
-    returns the clan object of given clan name
-    """
-    for clan in game.clan.all_other_clans:
-        if clan.name == clan_name:
-            return clan
 
 
 def create_new_cat(
@@ -1518,20 +1465,6 @@ def get_leader_life_notice() -> str:
     if game.clan.instructor.status.group == CatGroup.DARK_FOREST:
         return i18n.t("cat.history.leader_lives_left_df", count=game.clan.leader_lives)
     return i18n.t("cat.history.leader_lives_left_sc", count=game.clan.leader_lives)
-
-
-def get_other_clan_relation(relation):
-    """
-    converts int value into string relation and returns string: "hostile", "neutral", or "ally"
-    :param relation: the other_clan.relations value
-    """
-
-    if int(relation) >= 17:
-        return "ally"
-    elif 7 < int(relation) < 17:
-        return "neutral"
-    elif int(relation) <= 7:
-        return "hostile"
 
 
 def pronoun_repl(m, cat_pronouns_dict, raise_exception=False):
